validate using nose setUp and tearDown
